Enable indexing datasets by name
