+++ packages/syft/src/syft/serde/third_party.py
@@ -2,6 +2,7 @@
 from datetime import date
 from datetime import datetime
 from datetime import time
+import functools
 from io import BytesIO
 
 # third party
@@ -26,9 +27,12 @@ from result import Result
 import zmq.green as zmq
 
 # relative
+from ..types.tupledict import TupleDict
 from .deserialize import _deserialize as deserialize
+from .recursive_primitives import deserialize_kv
 from .recursive_primitives import recursive_serde_register
 from .recursive_primitives import recursive_serde_register_type
+from .recursive_primitives import serialize_kv
 from .serialize import _serialize as serialize
 
 recursive_serde_register(
@@ -122,6 +126,12 @@ recursive_serde_register(
     Timestamp,
     serialize=lambda x: serialize(x.value, to_bytes=True),
     deserialize=lambda x: Timestamp(deserialize(x, from_bytes=True)),
+)
+
+recursive_serde_register(
+    TupleDict,
+    serialize=serialize_kv,
+    deserialize=functools.partial(deserialize_kv, TupleDict),
 )
